refactor: consistently store abi metadata against contract class
