+++ src/abi-metadata.ts
@@ -1,9 +1,8 @@
-import type { Contract } from '@algorandfoundation/algorand-typescript'
 import type { AbiMethodConfig, BareMethodConfig, CreateOptions, OnCompleteActionStr } from '@algorandfoundation/algorand-typescript/arc4'
 import js_sha512 from 'js-sha512'
 import type { TypeInfo } from './encoders'
+import { Contract } from './impl/contract'
 import { getArc4TypeName as getArc4TypeNameForARC4Encoded } from './impl/encoded-types'
-import type { DeliberateAny } from './typescript-helpers'
 
 export interface AbiMetadata {
   methodName: string
@@ -14,40 +13,15 @@ export interface AbiMetadata {
   onCreate?: CreateOptions
   allowActions?: OnCompleteActionStr[]
 }
-const AbiMetaSymbol = Symbol('AbiMetadata')
-const overwrittenMetadata: Array<readonly [Contract | { new (): Contract }, Record<string, AbiMetadata>]> = []
+export const AbiMetaSymbol = Symbol('AbiMetadata')
 export const isContractProxy = Symbol('isContractProxy')
+const metadataStore: Map<{ new (): Contract }, Record<string, AbiMetadata>> = new Map()
 export const attachAbiMetadata = (contract: { new (): Contract }, methodName: string, metadata: AbiMetadata): void => {
-  const metadatas: Record<string, AbiMetadata> = (AbiMetaSymbol in contract ? contract[AbiMetaSymbol] : {}) as Record<string, AbiMetadata>
-
-  // classes inherited from the same parent shares the same metadata object stored in `AbiMetaSymbol`
-  // to prevent one subclass from overwriting the metadata of another subclass, we store the overwritten metadata in a separate array
-  if (metadatas[methodName]) {
-    let contractMetadatas = overwrittenMetadata.find(([c]) => c === contract)
-    if (!contractMetadatas) {
-      contractMetadatas = [contract, {}]
-      overwrittenMetadata.push(contractMetadatas)
-    }
-    contractMetadatas[1][methodName] = metadata
-  } else {
-    metadatas[methodName] = metadata
+  if (!metadataStore.has(contract)) {
+    metadataStore.set(contract, {})
   }
-  if (!(AbiMetaSymbol in contract)) {
-    Object.defineProperty(contract, AbiMetaSymbol, {
-      value: metadatas,
-      writable: true,
-      enumerable: true,
-    })
-  }
-}
-
-export const copyAbiMetadatas = <T extends Contract>(sourceContract: T, targetContract: T): void => {
-  const metadatas = getContractAbiMetadata(sourceContract)
-  Object.defineProperty(targetContract, AbiMetaSymbol, {
-    value: metadatas,
-    writable: true,
-    enumerable: false,
-  })
+  const metadatas: Record<string, AbiMetadata> = metadataStore.get(contract) as Record<string, AbiMetadata>
+  metadatas[methodName] = metadata
 }
 
 export const captureMethodConfig = <T extends Contract>(
@@ -61,23 +35,33 @@ export const captureMethodConfig = <T extends Contract>(
   metadata.allowActions = ([] as OnCompleteActionStr[]).concat(config?.allowActions ?? 'NoOp')
 }
 
-export const hasAbiMetadata = <T extends Contract>(contract: T): boolean => {
-  const contractClass = contract.constructor as { new (): T }
-  return (
-    Object.getOwnPropertySymbols(contractClass).some((s) => s.toString() === AbiMetaSymbol.toString()) || AbiMetaSymbol in contractClass
-  )
-}
-export const getContractAbiMetadata = <T extends Contract>(contract: T): Record<string, AbiMetadata> => {
-  const overwrittenMetadataEntry = overwrittenMetadata.find(([c]) => c === contract)
-  if ((contract as DeliberateAny)[AbiMetaSymbol]) {
-    return { ...((contract as DeliberateAny)[AbiMetaSymbol] as Record<string, AbiMetadata>), ...overwrittenMetadataEntry?.[1] }
+export const getContractAbiMetadata = <T extends Contract>(contract: T | { new (): T }): Record<string, AbiMetadata> => {
+  // Initialize result object to store merged metadata
+  const result: Record<string, AbiMetadata> = {}
+
+  // Get the contract's class
+  let currentClass = contract instanceof Contract ? (contract.constructor as { new (): T }) : contract
+
+  // Walk up the prototype chain
+  while (currentClass && currentClass.prototype) {
+    // Find metadata for current class
+    const currentMetadata = metadataStore.get(currentClass)
+
+    if (currentMetadata) {
+      // Merge metadata with existing result (don't override existing entries)
+      const classMetadata = currentMetadata
+      for (const [methodName, metadata] of Object.entries(classMetadata)) {
+        if (!(methodName in result)) {
+          result[methodName] = metadata
+        }
+      }
+    }
+
+    // Move up the prototype chain
+    currentClass = Object.getPrototypeOf(currentClass)
   }
-  const contractClass = contract.constructor as { new (): T }
-  const s = Object.getOwnPropertySymbols(contractClass).find((s) => s.toString() === AbiMetaSymbol.toString())
-  const metadatas: Record<string, AbiMetadata> = (
-    s ? (contractClass as DeliberateAny)[s] : AbiMetaSymbol in contractClass ? contractClass[AbiMetaSymbol] : {}
-  ) as Record<string, AbiMetadata>
-  return { ...metadatas, ...overwrittenMetadataEntry?.[1] }
+
+  return result
 }
 
 export const getContractMethodAbiMetadata = <T extends Contract>(contract: T, methodName: string): AbiMetadata => {
