+++ src/internal/arc4.ts
@@ -1,2 +1,3 @@
 export * from '@algorandfoundation/algorand-typescript/arc4'
-export { abimethod, baremethod, Contract, methodSelector } from '../impl/contract'
+export { abimethod, baremethod, Contract } from '../impl/contract'
+export { methodSelector } from '../impl/method-selector'
